+++ src/main/java/com/crejo/moviereviews/service/review/ReviewServiceImpl.java
@@ -31,7 +31,7 @@ public class ReviewServiceImpl implements ReviewService {
 
 
     @Override
-    public void addReview(String movieTitle, String userName, Double movieScore) {
+    public void addReview(String movieTitle, String userName, Integer movieScore) {
         userRepository.findUser(userName)
                 .ifPresentOrElse(user -> {
                     movieRepository.findMovie(movieTitle)
